backend/pyproject.toml
[project]
name = "chimera-backend"
version = "0.1.0"
description = "Chimera v4.0 pre-runner backend (paper-method pipeline)."
requires-python = ">=3.11"
readme = "README.md"
license = { file = "../LICENSE" }
authors = [{ name = "Ascot Wealth Management" }]

dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.4.0",
  "imbalanced-learn>=0.12.0",
  "joblib>=1.3.0",
  "google-cloud-storage>=2.16.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
